+++ stock_uom_dimension/models/stock.py
@@ -37,7 +37,7 @@ class StockMoveDimension(models.Model):
     _name = "stock.move.dimension"
 
     dimension_id = fields.Many2one('uom.dimension', required=True, ondelete='cascade')
-    quantity = fields.Integer('Quantité', required=True)
+    quantity = fields.Float('Quantité', required=True)
     stock_move_id = fields.Many2one('stock.move', required=True, ondelete='cascade')
     name = fields.Char(compute='get_name', store=True)
     display_name = fields.Char(compute='get_name', store=True)
